Add user story 12 & 16 and test files.
